Removed sqlite3 gem

+++ db/migrate/20151203090837_change_amount_to_double.rb
@@ -1,5 +1,5 @@
 class ChangeAmountToDouble < ActiveRecord::Migration
   def change
-      change_column :donations, :amount, :double, :default => 0.00
+      change_column :donations, :amount, :float, :default => 0.00
   end
 end
